+++ src/export.ts
@@ -1,4 +1,4 @@
-﻿export { parse } from './renderer/parser';
+﻿export { parse, escape } from './renderer/parser';
 export { bind } from './renderer/bind';
 export { render } from './renderer/render';
 export { caches } from './parser/cache';