+++ src/renderer/render/media/twitter.ts
@@ -1,5 +1,5 @@
 ﻿import DOM from 'typed-dom';
-import { parse } from '../../parser';
+import { parse, escape } from '../../parser';
 import { Cache } from 'spica/cache';
 import { sanitize } from 'dompurify';
 
@@ -39,7 +39,7 @@ export function twitter(url: string): HTMLElement | void {
         void document.head.appendChild(script);
       },
       error({ statusText }) {
-        outer.innerHTML = parse(`*${url}\\\n-> ${statusText}*`).querySelector('p')!.innerHTML;
+        outer.innerHTML = parse(`*${escape(url)}\\\n-> ${escape(statusText)}*`).querySelector('p')!.innerHTML;
       },
     });
     return outer;